requirements.txt
aiofiles
aiohttp
aiosqlite
asyncio
beautifulsoup4
cachetools
dirtyjson
fastapi
httpx
lxml
nest_asyncio
pandas
psutil
pydantic
pytz
requests
sqlalchemy
starlette
tabulate 
tdqm
timebudget
uvicorn[standard]
watchgod
zstandard
python-decouple
uvloop
pyyaml